pyproject.toml
[tool.poetry]
name = "capsim"
version = "0.1.0"
description = "Data Generator Simulation"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.0"
typer = "^0.9.0"
pyyaml = "^6.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
isort = "^5.0"
mypy = "^1.0"
faker = "^25.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 